Return immediately for empty URL
Found during the investigation for https://github.com/BrightspaceHypermediaComponents/activities/pull/564, if the `d2l-labs-attachment` component is created with an empty `attachment.url`, we end up still trying to fetch it from the the unfurl service. This is just a tiny check/safety to avoid that network request, which will 404.